+++ changelog.md
@@ -5,6 +5,7 @@
   - Dark Mode should now apply to GPX Import Report
   - Custom MessageBoxes should now respect hotkeys in certain cases
   - Added PNG and WEBP file support
+  - Added an auto-updater. It's a bit tricky to test but I think it works. Report bugs pls.
 
 - BUGS & FIXES:
   - Patched GPX Import Report re #105